+++ applications/tests/tests_adding_sites.py
@@ -10,7 +10,7 @@ from applications.models import (
 )
 from cases.enums import CaseTypeEnum
 from lite_content.lite_api.strings import ExternalLocations
-from organisations.factories import SiteFactory
+from organisations.tests.factories import SiteFactory
 from static.countries.helpers import get_country
 from static.statuses.libraries.get_case_status import get_case_status_by_status
 from test_helpers.clients import DataTestClient
