+++ organisations/tests/tests_organisations.py
@@ -3,15 +3,15 @@ from parameterized import parameterized
 from rest_framework import status
 from rest_framework.reverse import reverse
 
-from addresses.factories import ForeignAddressFactory
+from addresses.tests.factories import ForeignAddressFactory
 from conf.authentication import EXPORTER_USER_TOKEN_HEADER
 from conf.constants import Roles, GovPermissions
 from conf.helpers import date_to_drf_date
 from lite_content.lite_api.strings import Organisations
 from organisations.enums import OrganisationType, OrganisationStatus
-from organisations.factories import OrganisationFactory
+from organisations.tests.factories import OrganisationFactory
 from organisations.models import Organisation
-from organisations.providers import OrganisationProvider
+from organisations.tests.providers import OrganisationProvider
 from static.statuses.enums import CaseStatusEnum
 from static.statuses.libraries.get_case_status import get_case_status_by_status
 from test_helpers.clients import DataTestClient
